requirements.txt
pyserial>=3.5
bleak>=0.21.0
colorama>=0.4.6
